Split astar test in two (TestTrivialPathNoUturns use different config)
